feat(dotnet-checker): pass dotnet checker feature flag to plugin and refactor
